add fastclick + fix other errors

+++ src/js/deletepatient.js
@@ -1,14 +1,17 @@
 $$('.btn--delete').on('click', function() {
-	localStorage.clear();
-	emptyPatientInfo();
-    $$('.noPatient').show();
+        id = this.id;
 
-	id = this.id;
-	var deletePatient = 'deletePatient';
-	$$.post(connection, {
-        "type": deletePatient,
-        'id': id
-    }, function(data) {
-        	console.log(data);
+    myApp.confirm('Dette vil slette patient og alt tilhørende', function() {
+        localStorage.clear();
+        emptyPatientInfo();
+        $$('.noPatient').show();
+
+        var deletePatient = 'deletePatient';
+        $$.post(connection, {
+            "type": deletePatient,
+            'id': id
+        }, function(data) {
+            console.log(data);
+        });
     });
 });
